update migration for preview image bucket
