+++ packages/utility/src/index.ts
@@ -4,15 +4,16 @@ export * from './clone';
 export * from './color';
 export * from './defaults';
 export * from './defaults/canvas';
+export * from './email';
 export * from './errors';
 export * from './font';
-export * from './email';
 export * from './id';
 export * from './image';
 export * from './initials';
 export * from './math';
 export * from './null';
 export * from './path';
+export * from './screenshot';
 export * from './string';
 export * from './tailwind';
 export * from './time';
