+++ packages/utility/src/screenshot.ts
@@ -1,0 +1,3 @@
+export function getScreenshotPath(projectId: string) {
+    return `public/${projectId}.png`;
+}
